% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBuffer.R
\name{extractBuffer}
\alias{extractBuffer}
\title{Extract data from a buffer}
\usage{
extractBuffer(shpData, buffer = NULL, shpBounds = NULL, width = 5000)
}
\arguments{
\item{shpData}{Data to be extracted}

\item{buffer}{Shapefile with the buffer to extract the data from}

\item{shpBounds}{Shapefile from which the buffer will be created if a shapefile with
the buffer has not been provided.}

\item{width}{Width of the buffer}
}
\value{
A shapefile  contaning the intersection of the data and the buffer
}
\description{
This function extracts giver data from a buffer by performing an
intersection. A shapefile containing only the buffer can be provided in order to speed
up the process. Otherwise, a shapefile with the bounds from with to create the buffer
must be provided with the size of the buffer.
}

