% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageUtils.R
\name{i18nTextOutput}
\alias{i18nTextOutput}
\title{i18nTextOutput}
\usage{
i18nTextOutput(id, lang = NULL, data = NULL, inline = FALSE, class = "",
  ...)
}
\arguments{
\item{...}{}
}
\description{
Internationalized wrapper to shiny textOutput. Based on textOutput2
Use only to display text
}

