% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtils.R
\name{launchEvent}
\alias{launchEvent}
\title{Launch an event in a Shiny application. Use only when the application becomes
too complex for simple reactive values}
\usage{
launchEvent(type, userInfo)
}
\arguments{
\item{type}{type of event to launch}

\item{userInfo}{a ReactiveValues object containing all the information about the
application}
}
\description{
The function will create - or update if it already exists - a list called
event in the userInfo object. The list contains two elements: value which is just a
numerical counter automatically increased when an event is launched and on which
dependencies should be taken and type to describe the type of event launched
}

