% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languageUtils.R
\name{loadLanguages}
\alias{loadLanguages}
\title{Title}
\usage{
loadLanguages(path.to.file, reload = TRUE)
}
\arguments{
\item{path.to.file}{The path to the language file. The file must be a semi-colon
separated file with one column containing the keys and the others containing the
languages. The first line should contain the title for each column}

\item{reload}{Reload language list if the language file has been modified since last
access}
}
\value{
a data.frame containing all the language keys
}
\description{
Title
}

