% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUtils.R
\name{textOutput2}
\alias{textOutput2}
\title{textOutput2}
\usage{
textOutput2(id, content = NULL, container = if (inline) htmltools::span else
  htmltools::div, inline = FALSE, class = "", ...)
}
\arguments{
\item{...}{}
}
\description{
Alternative to shiny textOutput to allow multiple classes on the
text element and to allow the user of returning text directly
instead of having to wrap it in a renderText element before
}

